+++ src/robots/setup.py
@@ -4,7 +4,7 @@ with open("README.md", "r", encoding="utf-8") as fh:
     long_description = fh.read()
 
 setuptools.setup(
-    name="pybullet_robots",
+    name="robots",
     version="0.1",
     description="This package implements the base class for the PyBullet robots as well as the specific robots",
     long_description=long_description,
